requirements.txt
pandas
numpy
plotly
plotly-express
matplotlib
streamlit
black
openpyxl
requests
ElexonDataPortal
python-dotenv